Logging structure refactor
This splits up pkg/runtime/log into pkg/log, pkg/internal/log, and
pkg/log/zap.

- pkg/log contains generic logging helpers, and the generic root logger
  It *only* depends on logr, and thus won't pull in Zap if you don't
  need it.

- pkg/internal/log contains the internal log handle (with the
  appropriate name).  Nothing outside CR should be logging with this
  handle anyway, so this enforces it.

- pkg/log/zap contains Zap-related setup code, and thus CR only has a
  dependency on Zap if you pull in this package.

- pkg/runtime/log remains as a deprecated package to ease refactoring
  and compatibility.  While practically a few types have been removed
  from this package, in reality most consumers shouldn't notice.
