+++ pkg/log/null.go
